requirements.txt
alembic==1.8.1
aniso8601==9.0.1
apache-libcloud==3.6.1
arrow==1.2.3
atomicwrites==1.4.1
attrs==22.1.0
beautifulsoup4==4.11.1
bidict==0.22.0
blinker==1.5
boto3==1.26.105
cachelib==0.9.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
coverage==6.5.0
croniter==1.3.7
cryptography==36.0.0
cssmin==0.2.0
cssutils==2.6.0
DateTime==4.7
Deprecated==1.2.13
dnspython==2.2.1
elastic-transport==8.4.0
elasticsearch==7.13.2
email-validator==1.3.0
fasteners==0.18
flake8==4.0.1
Flask==2.0.2
Flask-Admin==1.6.0
Flask-Assets==2.0
Flask-Cache==0.13.1
Flask-Caching==2.0.1
Flask-Cloudy==1.1.0
Flask-Cors==3.0.10
Flask-Dance==5.1.0
Flask-DebugToolbar==0.11.0
Flask-Ext==0.1
Flask-Limiter==2.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-RESTful==0.3.9
Flask-RQ2==18.3
Flask-Script==2.0.6
Flask-SocketIO==5.3.1
Flask-SQLAlchemy==2.5.1
Flask-SQLAlchemy-Cache==0.1.5
Flask-Testing==0.8.1
Flask-WTF==1.0.0
geopy==2.3.0
googlemaps==4.10.0
greenlet==2.0.1
gunicorn==20.1.0
hashids==1.3.1
humanize==3.12.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsmin==3.0.0
limits==2.7.1
lockfile==0.12.2
Mako==1.2.3
MarkupSafe==2.1.1
mccabe==0.6.1
oauthlib==3.2.2
packaging==21.3
pandas==2.0.2
pluggy==1.0.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pygeocodio==1.2.0
PyJWT==1.7.1
PyMySQL==1.0.2
pynliner==0.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-engineio==4.3.4
python-slugify==6.1.2
python-socketio==5.7.2
pytz==2022.6
raven==6.10.0
redis==4.0.2
requests==2.26.0
requests-oauthlib==1.3.1
rq==1.10.0
rq-dashboard==0.6.1
rq-scheduler==0.11.0
six==1.16.0
snowflake-connector-python==3.0.2
soupsieve==2.3.2.post1
SQLAlchemy==1.4.27
SQLAlchemy-Utils==0.37.9
stripe==4.2.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
URLObject==2.4.3
webassets==2.0
Werkzeug==2.0.3
wrapt==1.14.1
WTForms==3.0.1
zipp==3.10.0
zope.interface==5.5.1
